+++ DESARROLLO/HITO1/ITERACION 3/ArquitecturaXComponentes/entitySystem/systems/moveSystem.h
@@ -22,7 +22,7 @@ public:
     moveSystem(const moveSystem& orig);
     moveSystem(entityManager*);
     virtual ~moveSystem();
-    void update(double);
+    void update(float);
 private:
 
 };
