+++ DESARROLLO/HITO1/ITERACION 3/ArquitecturaXComponentes/entitySystem/systems/system.h
@@ -23,7 +23,7 @@ public:
     system(const system& orig);
     system(entityManager *);
     virtual ~system();
-    void update(unsigned int);
+    void update(float);
     entityManager* getEntityManager();
 private:
     entityManager *eManager;
